Abstract EntityHeader to be used for both landing and detail views
