+++ packages/manager/src/components/EntityHeader/index.tsx
@@ -1,0 +1,1 @@
+export { default, HeaderProps } from './EntityHeader';
